feat: make MuiTabs customizable
