+++ packages/widget/src/components/Header/Header.style.ts
@@ -3,15 +3,12 @@ import {
   Avatar,
   Box,
   Button,
-  alpha,
   buttonClasses,
   styled,
-  tabsClasses,
 } from '@mui/material';
 import type { WidgetSubvariant } from '../../types/widget.js';
 import { getContrastAlphaColor } from '../../utils/colors.js';
 import { avatarMask12 } from '../Avatar/utils.js';
-import { Tabs } from '../Tabs/Tabs.style.js';
 
 export const HeaderAppBar = styled(AppBar)(({ theme }) => ({
   backgroundColor: 'transparent',
@@ -83,19 +80,6 @@ export const HeaderControlsContainer = styled(Box)(({ theme }) => ({
       marginRight: theme.spacing(-1.25),
     },
   }),
-}));
-
-export const SplitTabs = styled(Tabs)(({ theme }) => ({
-  backgroundColor:
-    theme.palette.mode === 'light'
-      ? alpha(theme.palette.common.black, 0.04)
-      : theme.palette.background.paper,
-  [`.${tabsClasses.indicator}`]: {
-    backgroundColor:
-      theme.palette.mode === 'light'
-        ? theme.palette.background.paper
-        : theme.palette.common.black,
-  },
 }));
 
 export const WalletAvatar = styled(Avatar)(({ theme }) => ({
