+++ packages/widget/src/components/Header/NavigationTabs.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'react-i18next';
 import { useFieldActions } from '../../stores/form/useFieldActions.js';
 import { useSplitSubvariantStore } from '../../stores/settings/useSplitSubvariantStore.js';
-import { Tab } from '../Tabs/Tabs.style.js';
-import { HeaderAppBar, SplitTabs } from './Header.style.js';
+import { Tab, Tabs } from '../Tabs/Tabs.style.js';
+import { HeaderAppBar } from './Header.style.js';
 
 export const NavigationTabs = () => {
   const { t } = useTranslation();
@@ -21,7 +21,7 @@ export const NavigationTabs = () => {
 
   return (
     <HeaderAppBar elevation={0} sx={{ paddingTop: 1, paddingBottom: 0.5 }}>
-      <SplitTabs
+      <Tabs
         value={state === 'swap' ? 0 : 1}
         onChange={handleChange}
         aria-label="tabs"
@@ -29,7 +29,7 @@ export const NavigationTabs = () => {
       >
         <Tab label={t('header.swap')} disableRipple />
         <Tab label={t('header.bridge')} disableRipple />
-      </SplitTabs>
+      </Tabs>
     </HeaderAppBar>
   );
 };
