+++ packages/widget/src/utils/colors.ts
@@ -15,8 +15,3 @@ export const getInfoBackgroundColor = (theme: Theme) =>
   theme.palette.mode === 'light'
     ? alpha(theme.palette.info.main, 0.12)
     : alpha(theme.palette.info.main, 0.16);
-
-export const getCardFieldsetBackgroundColor = (theme: Theme) =>
-  theme.palette.mode === 'light'
-    ? alpha(theme.palette.common.black, 0.04)
-    : theme.palette.grey[800];
